requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic[email]
bcrypt
passlib
python-jose
# python-multipart
python-dotenv
pymysql
python-multipart    # Form data